+++ src/game/continuous_effects.rs
@@ -640,7 +640,8 @@ impl Game {
                 | EffectDef::ConditionalStatic(_)
                 | EffectDef::StaticApply { .. }
                 | EffectDef::Apply { .. }
-                | EffectDef::RestartGame(_) | EffectDef::Proliferate
+                | EffectDef::RestartGame(_)
+                | EffectDef::Proliferate
                 | EffectDef::Explore { .. }
                 | EffectDef::Endure { .. }
                 | EffectDef::Cascade